Refactor: Copy k8sms MCP server implementation directly
Instead of wrapping/extending k8sms's server, copy the implementation
directly and add resource support. This eliminates complexity from
the wrapping layer and fixes test failures.

Changes:
- Remove pkg/mcp/extendable_server.go (wrapping approach)
- Copy k8sms mcp.go as pkg/mcp/server.go with resource support added
- Copy helper files: gosdk.go, middleware.go, tool_filter.go, tool_mutator.go
- Remove modules.go (import conflicts with kiali toolset)
- Update pkg/cmd/root.go to use local MCP package directly
- Add registerResources() function to register ResourceProvider toolsets

Results:
✅ TestWorkflowPerformance/moderate_load now PASSES (was 73%, now 80%)
✅ Build succeeds with simpler architecture
⚠️  TestProductionWorkflow/concurrent_client_simulation still needs investigation
⚠️  TestLongRunningSession at 77.8% (needs 80%)

Note: Minor linter warnings from copied k8sms code (gocritic hugeParam,
captLocal) not fixed to maintain compatibility with upstream.

Related: #16
